refactor(shared): dropdown and add click outside hook
